build: use gotestsum
This makes the output more readable, notably if we want xUnit output.
This allows the CI to only run `make test-race` for example and get
both a nice output and xUnit XML file. This is somewhat important as
if we run just `test-xml`, the output is difficult to read. If we
first use `test` then `test-xml`, if `test` fails, we won't get the
xUnit file.